requirements.txt
flask~=3.0.3
requests~=2.32.3
gunicorn
mysql-connector-python~=9.1.0
python-dotenv~=1.0.1
pymongo~=4.10.1